add simple CPU implementation
